let clang 6.0.1 compile

+++ src/finite_fields/gf2n/gf2nin.cc
@@ -80,7 +80,7 @@ std::istream & operator >> (std::istream & in_strm, gf2n & a)
 	}
 	else {
 		str = line+i+1;
-		line[i] = (char) NULL;
+		line[i] = (char)(size_t) NULL;
 
 		if (strcmp(line, "HEX") == 0) {
 			hexin(a, str);
